+++ src/statemanager.hpp
@@ -30,11 +30,11 @@ protected:
 
   virtual void on_keyup(const input::keyevent &event) noexcept;
 
-  virtual void on_joystickbuttondown(int who, const input::joystickevent &event) noexcept;
+  virtual void on_joystickbuttondown(uint8_t who, const input::joystickevent &event) noexcept;
 
-  virtual void on_joystickbuttonup(int who, const input::joystickevent &event) noexcept;
+  virtual void on_joystickbuttonup(uint8_t who, const input::joystickevent &event) noexcept;
 
-  virtual void on_joystickaxismotion(int who, const input::joystickaxisevent &event) noexcept;
+  virtual void on_joystickaxismotion(uint8_t who, const input::joystickaxisevent &event) noexcept;
 
   virtual void on_collision(const collisionevent &event) noexcept;
